requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0

# Vision
opencv-python>=4.8.0
mediapipe>=0.10.0

# Generator
torch>=2.0.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
